% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pkgpear}
\alias{install_pkgpear}
\title{Install \emph{*pear} package}
\usage{
install_pkgpear(pkg, login, pass = NULL, quiet = FALSE)
}
\arguments{
\item{pkg}{package name}

\item{login}{bitbucket login.}

\item{pass}{bitbucket password. You should not put your password directly. Read details.}

\item{quiet}{if TRUE suppresses output from this function.}
}
\description{
Function which allows you to install \emph{*pear} package.
}
\details{
Regarding \emph{pass} argument, ideally, use default value: \emph{pass = NULL}. After that,
the popup window shows up which ask you for your bitbucket password. In that case, your password
will not be stored in memory. Alternatively, you can use your local .Renviron file.
}
