% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonverse.R
\docType{package}
\name{pearsonverse-package}
\alias{pearsonverse}
\alias{pearsonverse-package}
\title{pearsonverse: A package which stores all pearson-related packages}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.pearson.com/projects/EF/repos/pearsonverse/browse}
}

}
\keyword{internal}
