% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearsonverse.R
\docType{package}
\name{pearsonverse-package}
\alias{pearsonverse}
\alias{pearsonverse-package}
\title{pearsonverse: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
}
\seealso{
Useful links:
\itemize{
  \item \url{}
}

}
\keyword{internal}
